Add files via upload
This code allows the user to scan a RFID tag and allocate a number to it, which is then built into a csv file for later use. I use this to assign a number to my RFID tags in a holder that each have a number assigned, so I know which holder number relates to which EPC. No writing to the RFID tag itself occurs, only an allocation table is created.